# agent-skill/tools/assess_dog_health.openapi.yaml
openapi: 3.1.0
info:
  title: Dog Happiness Assessment and Care Consultation API
  version: 2.3.0-beta
  description: Generates a reproducible dog happiness self-check and evidence-bounded feeding consultation. Agents must never score, add advice from memory, or override safety results.
servers:
  - url: https://dog-health-check.vercel.app
    description: Production Vercel server
paths:
  /api/assess:
    post:
      operationId: assessDogHappiness
      summary: Generate a dog happiness self-check dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessRequest'
      responses:
        '200':
          description: Complete evidence-informed dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessResponse'
        '400':
          description: Invalid or incomplete input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/consult:
    post:
      operationId: consultDogCareKnowledge
      summary: Get evidence-bounded dog feeding and wellbeing guidance after assessment
      description: Re-runs the assessment safety rules, retrieves only local knowledge-base guidance, and returns the sources actually used. Safety responses never include an ordinary feeding plan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultRequest'
      responses:
        '200':
          description: Cited guidance or a safety-first professional support route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultResponse'
        '400':
          description: Invalid assessment or consultation question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    AssessRequest:
      type: object
      additionalProperties: false
      required: [age, breed, size, bodyCondition, foodTypes, snackLevel, movement, recentSignals, homeEnvironment, mentalState]
      properties:
        dogName: { type: string, maxLength: 60 }
        age: { type: string, minLength: 1, maxLength: 60 }
        breed: { type: string, minLength: 1, maxLength: 80 }
        weight: { type: string, maxLength: 80 }
        size:
          type: string
          enum: [small, medium, large, unknown]
        bodyCondition:
          type: string
          enum: [thin, ideal, slightly_fat, obese, unknown]
        foodTypes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [kibble, canned, freeze_dried, fresh, homemade, mixed]
        snackLevel:
          type: string
          enum: [low, medium, high]
        movement:
          type: object
          additionalProperties: false
          required: [outdoorFrequency, dailyMinutes, sniffing]
          properties:
            outdoorFrequency:
              type: string
              enum: [rare, once_toilet_only, one_to_two, two_to_three, two_plus]
            dailyMinutes:
              type: string
              enum: [unknown, under_10, 10_20, around_30, over_30]
            sniffing:
              type: string
              enum: [none, little, normal, rich, unknown]
        recentSignals:
          type: array
          minItems: 1
          uniqueItems: true
          description: normal must be selected alone. Persistent observations and urgent warning signals use the same array; the API alone determines safety routing.
          items:
            type: string
            enum: [normal, unstable_poop, bad_breath, paw_licking_or_scratching, persistent_cough, drinking_or_urination_change, unexplained_weight_change, low_energy, appetite_change, vomiting, diarrhea, pain, mobility_issue, breathing_difficulty, collapse_seizure_or_fainting, swollen_abdomen_or_unproductive_retching]
        homeEnvironment:
          type: array
          minItems: 1
          uniqueItems: true
          description: none and unknown must each be selected alone.
          items:
            type: string
            enum: [aromatherapy, strong_cleaner, secondhand_smoke, plastic_bowl, pesticide, none, unknown]
        mentalState:
          type: object
          additionalProperties: false
          required: [positiveEngagement, relaxation, socialConnection, distressSignals]
          properties:
            positiveEngagement:
              type: string
              enum: [often, sometimes, rare, unknown]
            relaxation:
              type: string
              enum: [easy, sometimes_difficult, often_difficult, unknown]
            socialConnection:
              type: string
              enum: [secure, variable, withdrawn, unknown]
            distressSignals:
              type: array
              minItems: 1
              uniqueItems: true
              description: none and unknown must each be selected alone.
              items:
                type: string
                enum: [frequent_fear_or_hiding, separation_distress, persistent_pacing_or_vocalizing, repetitive_behavior, sudden_behavior_change, aggression_safety_risk, self_injury, none, unknown]
    AssessResponse:
      type: object
      additionalProperties: false
      required: [status, statusText, coreConclusion, happinessScore, healthScore, mentalWellbeingScore, assessmentConfidence, answeredCoverage, healthDimensionScores, mentalDimensionScores, strengths, coreRisk, todayAction, supportRoute, supportReminder, allowShare, shareCopy, knowledgeVersion, evidenceRefs, dashboardReport, shareGuide]
      properties:
        status:
          type: string
          enum: [excellent, good, watch, adjust, risk_confirm, insufficient, behavior_support, red_vet]
        statusText: { type: string }
        coreConclusion: { type: string }
        happinessScore: { type: [integer, "null"], minimum: 0, maximum: 100 }
        healthScore: { type: [integer, "null"], minimum: 0, maximum: 100 }
        mentalWellbeingScore: { type: [integer, "null"], minimum: 0, maximum: 100 }
        assessmentConfidence:
          type: string
          enum: [high, medium, insufficient]
        answeredCoverage: { type: number, minimum: 0, maximum: 1 }
        healthDimensionScores:
          type: object
          additionalProperties: false
          required: [body, diet, movement, recent, environment]
          properties:
            body: { type: [integer, "null"], minimum: 0, maximum: 100 }
            diet: { type: [integer, "null"], minimum: 0, maximum: 100 }
            movement: { type: [integer, "null"], minimum: 0, maximum: 100 }
            recent: { type: [integer, "null"], minimum: 0, maximum: 100 }
            environment: { type: [integer, "null"], minimum: 0, maximum: 100 }
        mentalDimensionScores:
          type: object
          additionalProperties: false
          required: [positiveEngagement, relaxation, socialConnection, distress]
          properties:
            positiveEngagement: { type: [integer, "null"], minimum: 0, maximum: 100 }
            relaxation: { type: [integer, "null"], minimum: 0, maximum: 100 }
            socialConnection: { type: [integer, "null"], minimum: 0, maximum: 100 }
            distress: { type: [integer, "null"], minimum: 0, maximum: 100 }
        strengths:
          type: array
          minItems: 1
          items: { type: string }
        coreRisk:
          type: object
          additionalProperties: false
          required: [title, reason, domain]
          properties:
            title: { type: string }
            reason: { type: string }
            domain:
              type: string
              enum: [health, mental, safety, information]
        todayAction:
          type: object
          additionalProperties: false
          required: [title, body]
          properties:
            title: { type: string }
            body: { type: string }
        supportRoute:
          type: string
          enum: [none, vet, veterinary_behavior]
        supportReminder: { type: string }
        allowShare: { type: boolean }
        shareCopy:
          type: string
          description: Must be empty for insufficient, behavior_support, and red_vet.
        knowledgeVersion: { type: string }
        evidenceRefs:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [id, title, organization, level, url]
            properties:
              id: { type: string }
              title: { type: string }
              organization: { type: string }
              level:
                type: string
                enum: [A1, A2, B, C]
              url: { type: string, format: uri }
        dashboardReport:
          type: object
          required: [title, subtitle, scoreLabel, sections]
          properties:
            title: { type: string }
            subtitle: { type: string }
            scoreLabel: { type: string }
            sections:
              type: array
              items:
                type: object
                required: [id, title]
                properties:
                  id: { type: string }
                  title: { type: string }
                  body: { type: string }
                  items: { type: array, items: { type: string } }
        shareGuide:
          type: object
          required: [primaryCta, safetyNote, actions]
          properties:
            primaryCta: { type: string }
            safetyNote: { type: string }
            actions:
              type: array
              items:
                type: object
                required: [id, label, description, enabled]
                properties:
                  id:
                    type: string
                    enum: [copy_text, download_image, wechat_moments, xiaohongshu, weibo]
                  label: { type: string }
                  description: { type: string }
                  enabled: { type: boolean }
    ConsultRequest:
      type: object
      additionalProperties: false
      required: [assessment, message]
      properties:
        assessment:
          $ref: '#/components/schemas/AssessRequest'
        message:
          type: string
          minLength: 1
          maxLength: 1000
    ConsultResponse:
      type: object
      additionalProperties: false
      required: [mode, topic, title, answer, plan, watchFor, citations, followUpPrompts, safetyRoute, knowledgeVersion, disclaimer]
      properties:
        mode:
          type: string
          enum: [guidance, safety]
        topic:
          type: string
          enum: [daily_feeding, weight_management, treats, diet_transition, home_prepared, raw_food, feeding_enrichment, life_stage, safety]
        title: { type: string }
        answer: { type: string }
        plan:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [day, action]
            properties:
              day: { type: string }
              action: { type: string }
        watchFor:
          type: array
          items: { type: string }
        citations:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [id, title, organization, level, url]
            properties:
              id: { type: string }
              title: { type: string }
              organization: { type: string }
              level:
                type: string
                enum: [A1, A2, B, C]
              url: { type: string, format: uri }
        followUpPrompts:
          type: array
          items: { type: string }
        safetyRoute:
          type: string
          enum: [none, vet, veterinary_behavior]
        knowledgeVersion: { type: string }
        disclaimer: { type: string }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, details]
          properties:
            message: { type: string }
            details:
              type: array
              items: { type: string }